+++ base/dates/io.jl
@@ -28,10 +28,10 @@ function tryparsenext end
 """
     format(io::IO, tok::AbstractDateToken, dt::TimeType, locale)
 
-format the `tok` token from `dt` and write it to `io`. The formatting can
+Format the `tok` token from `dt` and write it to `io`. The formatting can
 be based on `locale`.
 
-all subtypes of `AbstractDateToken` must define this method in order
+All subtypes of `AbstractDateToken` must define this method in order
 to be able to print a Date / DateTime object according to a `DateFormat`
 containing that token.
 """